subida de clase  actividad

+++ Entrega.java
@@ -2,16 +2,16 @@ import java.util.Comparator;
 public class Entrega{
     private String nombre;
     private String comentarios;
-    private int nota;
+    private double nota;
     private String fechaEntrega;
 
-    public Entrega(String nombre, String comentarios, int nota) {
+    public Entrega(String nombre, String comentarios, double nota) {
         this.nombre = nombre;
         this.comentarios = comentarios;
         this.nota = nota;
     }
 
-    public int getNota() { return nota;}
+    public double getNota() { return nota;}
     public String getNombre() {return nombre;}
     public String getComentarios() { return comentarios;}
 }
